+++ src/scorers/Scorer.java
@@ -3,5 +3,6 @@ package scorers;
 import crawler.Page;
 
 public interface Scorer {
-	int getScore(Page page);
+	double getScore(Page page);
+	
 }
